backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
pydantic
python-dotenv
alembic
pytest